ASD: increase defrag percentage (#29)
After the checkpoint is flushed, we go through list of Keys modified
in the checkpoint and issue a delete for all of these keys. The block
is marked for defrag, once the amount of usable data in ssd_write_buf*
decrases beyond some point. If the defrag percet is too low, probablity
of deleting a Key, that is just defragged incrases significantly.

The defrag increase the load on the storage devices. It can quickly
fill the write queue. ASD starts giving OVERLOAD errors, once the write
queue grow beyond certain limit.

Signed-off-by: Prasad Joshi <Prasad.Joshi@primaryio.com>